+++ server/routes/optup.js
@@ -4,18 +4,33 @@ const app = express();
 let cors = require("cors");
 const optup = require('../models/optup');
 router.use(cors());
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
 
 
 
+let authenticate=(req,res,next)=>{
+    let token=req.header('x-access-token');
+    jwt.verify(token,User.getJWTSecret(),(err,decoded)=>{
+        if(err){
+            res.status(401).send(err);
+        }
+        else{
+            req.user_id=decoded._id;
+            next(); 
+        }
 
-router.post('/add', async(req, res, next) => {  
+    });
+}
+
+router.post('/add', authenticate, async(req, res, next) => {  
   
     // console.log(req.body);
     const {idenseignant,nomenseignant,up,creneaux,periode} = req.body;
     
     try {   
             const addoptup = new optup({
-                idenseignant,nomenseignant,up,creneaux,periode});
+                idenseignant,nomenseignant,up,creneaux,periode,_userId:req.user_id });
     
             await addoptup.save();
             res.status(201).json(addoptup);
@@ -28,8 +43,10 @@ router.post('/add', async(req, res, next) => {
 
 
     
-router.get("/read", async(req, res) => {
-    optup.find({}, (err, result) => {
+router.get("/read",authenticate, async(req, res) => {
+    optup.find({
+        _userId:req.user_id
+    }, (err, result) => {
   
         if (err) {
             res.send(err)
@@ -44,10 +61,10 @@ router.get("/read", async(req, res) => {
 
   
    
-  router.delete('/:id', async(req, res) => {
+  router.delete('/:id',authenticate,async(req, res) => {
 
     const id = req.params.id;
-    await optup.findByIdAndRemove(id).exec();
+    await optup.findByIdAndRemove({_id:id,user_id:req.user_id}).exec();
     res.send("deleted");
 
 
@@ -58,10 +75,10 @@ router.get("/read", async(req, res) => {
 
 
 
-router.put("/update/:id", async(req, res) => {
+router.put("/update/:id",authenticate,  async(req, res) => {
     try {
         const { id } = req.params;
-        const updatecomposant = await enseignant.findByIdAndUpdate(id, req.body, {
+        const updatecomposant = await enseignant.findOneAndUpdate({_id:id,user_id:req.user_id}, req.body, {
             new: true
         });
