+++ server/routes/type.js
@@ -4,18 +4,32 @@ const app = express();
 let cors = require("cors");
 const type = require('../models/type');
 router.use(cors());
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
+
+let authenticate=(req,res,next)=>{
+    let token=req.header('x-access-token');
+    jwt.verify(token,User.getJWTSecret(),(err,decoded)=>{
+        if(err){
+            res.status(401).send(err);
+        }
+        else{
+            req.user_id=decoded._id;
+            next(); 
+        }
+
+    });
+}
 
 
-
-
-router.post('/add', async(req, res, next) => {  
+router.post('/add', authenticate, async(req, res, next) => {  
   
     // console.log(req.body);
     const {typeenseignement,nbreheures} = req.body;
     
     try {   
             const addtype = new type({
-                typeenseignement,nbreheures});
+                typeenseignement,nbreheures,_userId:req.user_id });
     
             await addtype.save();
             res.status(201).json(addtype);
@@ -32,8 +46,10 @@ router.post('/add', async(req, res, next) => {
 
 
     
-router.get("/read", async(req, res) => {
-    type.find({}, (err, result) => {
+router.get("/read",authenticate, async(req, res) => {
+    type.find({
+        _userId:req.user_id
+    }, (err, result) => {
   
         if (err) {
             res.send(err)
@@ -55,10 +71,10 @@ router.get("/read", async(req, res) => {
 
   
    
-  router.delete('/:id', async(req, res) => {
+  router.delete('/:id',authenticate, async(req, res) => {
 
     const id = req.params.id;
-    await type.findByIdAndRemove(id).exec();
+    await type.findByIdAndRemove({_id:id,user_id:req.user_id}).exec();
     res.send("deleted");
 
 
@@ -69,10 +85,10 @@ router.get("/read", async(req, res) => {
 
 
 
-router.put("/update/:id", async(req, res) => {
+router.put("/update/:id",authenticate, async(req, res) => {
     try {
         const { id } = req.params;
-        const updatecomposant = await type.findByIdAndUpdate(id, req.body, {
+        const updatecomposant = await type.findOneAndUpdate({_id:id,user_id:req.user_id}, req.body, {
             new: true
         });
