+++ server/routes/up.js
@@ -4,19 +4,35 @@ const app = express();
 let cors = require("cors");
 const up = require('../models/up');
 router.use(cors());
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
+let authenticate=(req,res,next)=>{
+    let token=req.header('x-access-token');
+    jwt.verify(token,User.getJWTSecret(),(err,decoded)=>{
+        if(err){
+            res.status(401).send(err);
+        }
+        else{
+            req.user_id=decoded._id;
+            next(); 
+        }
+
+    });
+}
 
 
 
 
 
-router.post('/add', async(req, res, next) => {  
+
+router.post('/add', authenticate, async(req, res, next) => {  
   
     // console.log(req.body);
     const {nomup,nomdepartement} = req.body;
     
     try {   
             const addclasse = new up    ({
-                nomup,nomdepartement});
+                nomup,nomdepartement,_userId:req.user_id  });
     
             await addclasse.save();
             res.status(201).json(addclasse);
@@ -30,8 +46,10 @@ router.post('/add', async(req, res, next) => {
 
 
     
-router.get("/read", async(req, res) => {
-    up.find({}, (err, result) => {
+router.get("/read",authenticate,  async(req, res) => {
+    up.find({
+        _userId:req.user_id
+    }, (err, result) => {
   
         if (err) {
             res.send(err)
@@ -46,10 +64,10 @@ router.get("/read", async(req, res) => {
 
  
    
-  router.delete('/:id', async(req, res) => {
+  router.delete('/:id',authenticate, async(req, res) => {
 
     const id = req.params.id;
-    await up.findByIdAndRemove(id).exec();
+    await up.findByIdAndRemove({_id:id,user_id:req.user_id}).exec();
     res.send("deleted");
 
 
@@ -60,10 +78,10 @@ router.get("/read", async(req, res) => {
 
 
 
-router.put("/update/:id", async(req, res) => {
+router.put("/update/:id",authenticate,  async(req, res) => {
     try {
         const { id } = req.params;
-        const updatecomposant = await up.findByIdAndUpdate(id, req.body, {
+        const updatecomposant = await up.findOneAndUpdate({_id:id,user_id:req.user_id}, req.body, {
             new: true
         });
